1.修复普通用户不能查看操作日志bug 2.新增用户协议 3.做申请消息缓存处理 4.修复钥匙管理相关bug

+++ module_lock/src/main/java/com/ut/module_lock/viewmodel/LockListFragVM.java
@@ -90,6 +90,7 @@ public class LockListFragVM extends AndroidViewModel {
                 .observeOn(Schedulers.io())
                 .subscribe(lockGroupResult -> {
                     List<LockGroup> list = lockGroupResult.getData();
+                    if (list == null) return;
                     LockGroup[] lockGroups = new LockGroup[list.size()];
                     if (isReset) {//刷新时清除所有组数据
                         LockKeyDaoImpl.get().deleteAll();
